Reduces complexity of UIKit Mocks
