+++ SourcingTests/UICollectionViewTableViewMocking.swift
@@ -1,0 +1,41 @@
+//
+//  Copyright (C) 2017 Lukas Schmidt.
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a 
+//  copy of this software and associated documentation files (the "Software"), 
+//  to deal in the Software without restriction, including without limitation 
+//  the rights to use, copy, modify, merge, publish, distribute, sublicense, 
+//  and/or sell copies of the Software, and to permit persons to whom the 
+//  Software is furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in 
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL 
+//  THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER 
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING 
+//  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
+//  DEALINGS IN THE SOFTWARE.
+//
+
+import UIKit
+
+protocol UICollectionViewTableViewMocking: class {
+    var lastUsedReuseIdentifiers: [String] { get set }
+    var cellMocks: [String: AnyObject] { get }
+}
+
+extension UICollectionViewTableViewMocking {
+    
+    func dequeueWithIdentifier<Cell>(_ identifier: String, forIndexPath indexPath: IndexPath) -> Cell {
+        lastUsedReuseIdentifiers.append(identifier)
+        
+        guard let cell = cellMocks[identifier] as? Cell else {
+            fatalError("Could not find cell mock with identifier: \(identifier)")
+        }
+        
+        return cell
+    }
+}
